单独查询模版属性数量接口不鉴权
