+++ src/scene_server/topo_server/service/field_template/service.go
@@ -64,6 +64,8 @@ func InitFieldTemplate(c *capability.Capability) {
 	// field template attribute
 	c.Utility.AddHandler(rest.Action{Verb: http.MethodPost, Path: "/findmany/field_template/attribute",
 		Handler: s.ListFieldTemplateAttr})
+	c.Utility.AddHandler(rest.Action{Verb: http.MethodPost, Path: "/findmany/field_template/attribute/count",
+		Handler: s.CountFieldTemplateAttr})
 
 	// field template unique
 	c.Utility.AddHandler(rest.Action{Verb: http.MethodPost, Path: "/findmany/field_template/unique",
